reorg(3/5): rename logging/ → observability/ + move events-emitter
Renames the folder to match modern industry terminology (OpenTelemetry,
Datadog, Honeycomb all talk about "observability" as the umbrella for
logs + metrics + traces + events). The folder now accurately groups
every file that describes "what the system is doing":

- `context.ts`, `error.ts`, `request.ts`, `response.ts` — existing
  log formatters (unchanged content)
- `events-emitter.ts` — the `RequestLifecycleEmitter`, moved from
  `src/services/`. Lifecycle events with correlation IDs are
  proto-traces, exactly an observability concern.

Classic and Home clients now import `RequestLifecycleEmitter` directly
from `../observability/index.ts` (removes the temporary bridge that
was in `resilience/index.ts` during commit 2).

Renames `tests/unit/logging.test.ts` → `tests/unit/observability.test.ts`
for symmetry with the source folder.

460 tests passing, 100 % coverage preserved.
