+++ src/observability/error.ts
