+++ src/observability/index.ts
@@ -1,5 +1,6 @@
 export type { LoggableRequestConfig } from './context.ts'
 
 export { createAPICallErrorData } from './error.ts'
+export { RequestLifecycleEmitter } from './events-emitter.ts'
 export { APICallRequestData } from './request.ts'
 export { APICallResponseData } from './response.ts'
