+++ src/observability/response.ts
