+++ src/resilience/index.ts
@@ -20,9 +20,3 @@ export {
   RateLimitError,
   TransientServerError,
 } from '../errors.ts'
-/*
- * Temporary bridge: `RequestLifecycleEmitter` still lives in services/
- * until commit 3 moves it to observability/. Re-exported here for a
- * single commit to keep `api.ts`/`home-api.ts` imports stable.
- */
-export { RequestLifecycleEmitter } from '../services/events-emitter.ts'
